Release/4.10.0 (#230)

+++ BrainCloudClient/Assets/Tests/PlayMode/TestFile.cs
@@ -13,8 +13,10 @@ using UnityEngine.TestTools;
 namespace Tests.PlayMode
 {
     public class TestFile : TestFixtureBase
-    {
+    {        
         private string _cloudPath = "";
+        private bool _shouldDeleteFiles = true;
+
         private int _returnCount = 0;
         private int _successCount = 0;
         private int _failCount = 0;
@@ -26,16 +28,22 @@ namespace Tests.PlayMode
         public override void TearDown()
         {
             _tc.bcWrapper.Client.DeregisterFileUploadCallbacks();
-            _tc.StartCoroutine(DeleteAllFiles());
+            if (_shouldDeleteFiles)
+            {
+                _tc.StartCoroutine(DeleteAllFiles());    
+            }
             base.TearDown();
             _returnCount = 0;
             _successCount = 0;
             _failCount = 0;
+            _shouldDeleteFiles = true;
         }
         
         [UnityTest]
         public IEnumerator TestListUserFiles()
         {
+            yield return _tc.StartCoroutine(_tc.SetUpNewUser(Users.UserA));
+            
             _tc.bcWrapper.FileService.ListUserFiles(_tc.ApiSuccess, _tc.ApiError);
             
             yield return _tc.StartCoroutine(_tc.Run());
@@ -43,6 +51,7 @@ namespace Tests.PlayMode
             _tc.bcWrapper.PlayerStateService.DeleteUser(_tc.ApiSuccess, _tc.ApiError);
             
             yield return _tc.StartCoroutine(_tc.Run());
+            LogResults("failed to list user files",_tc.successCount == 2);
         }
 
         [UnityTest]
@@ -99,8 +108,7 @@ namespace Tests.PlayMode
                 );
             yield return _tc.StartCoroutine(_tc.Run());
             yield return WaitForReturn(new[] { GetUploadId(_tc.m_response) });
-            
-            Assert.IsFalse(_failCount > 0);
+            LogResults("failed to upload file",_tc.successCount == 1);
         }
 
         [UnityTest]
@@ -142,12 +150,11 @@ namespace Tests.PlayMode
             yield return _tc.StartCoroutine(_tc.Run());
             
             yield return WaitForReturn(new[] { GetUploadId(_tc.m_response) });
-            
-            Assert.IsFalse(_failCount > 0);
+            LogResults("Couldn't upload all files",_tc.successCount == 2);
         }
 
         [UnityTest]
-        public IEnumerator TestSimpleUploadFailedFromPrivacySettings()
+        public IEnumerator TestSimpleUploadThenDownloadFromPrivacySettings()
         {
             yield return _tc.StartCoroutine(_tc.SetUpNewUser(Users.UserA));
             
@@ -169,8 +176,6 @@ namespace Tests.PlayMode
             
             yield return WaitForReturn(new[] { GetUploadId(_tc.m_response) });
 
-            Debug.Log($"Fail Count {_failCount}");
-            Assert.IsFalse(_failCount > 0);
 
             yield return new WaitForSeconds(2);
             
@@ -187,8 +192,8 @@ namespace Tests.PlayMode
                 Debug.Log(e);
                 _failCount++;
             }
-
-            Assert.IsFalse(_failCount == 0);
+            
+            LogResults("failed, it could be upload or download that failed.",_tc.successCount == 1);
         }
 
         [UnityTest]
@@ -222,15 +227,21 @@ namespace Tests.PlayMode
 
             var fileList = ((object[])((Dictionary<string, object>)_tc.m_response["data"])["fileList"]);
 
-            Debug.Log("\nDid fail = " + (_failCount > 0));
-            Debug.Log("File list length  = " + fileList.Length + "\n");
-            
-            Assert.IsFalse(_failCount <= 0 || fileList.Length > 0);
+            if (_failCount > 0)
+            {
+                Debug.Log("Number of failed callbacks: " + _failCount);
+                LogResults("",true);
+            }
+            else if (fileList.Length > 0)
+            {
+                LogResults("Response indicates the file did not get cancelled and successfully uploaded.",false);
+            }
         }
 
         [UnityTest]
         public IEnumerator TestDeleteUserFile()
         {
+            _shouldDeleteFiles = false;
             yield return _tc.StartCoroutine(_tc.SetUpNewUser(Users.UserA));
             
             _tc.bcWrapper.Client.RegisterFileUploadCallbacks(FileCallbackSuccess, FileCallbackFail);
@@ -262,11 +273,13 @@ namespace Tests.PlayMode
                 );
             
             yield return _tc.StartCoroutine(_tc.Run());
+            LogResults("failed to delete user files, check json response",_tc.successCount == 2);
         }
 
         [UnityTest]
         public IEnumerator TestDeleteUserFiles()
         {
+            _shouldDeleteFiles = false;
             yield return _tc.StartCoroutine(_tc.SetUpNewUser(Users.UserA));
             
             _tc.bcWrapper.Client.RegisterFileUploadCallbacks(FileCallbackSuccess, FileCallbackFail);
@@ -314,6 +327,8 @@ namespace Tests.PlayMode
                 );
             
             yield return _tc.StartCoroutine(_tc.Run());
+
+            LogResults("failed to delete all files OR uploading files failed.", _tc.successCount == 3);
         }
 
         [UnityTest]
@@ -351,6 +366,8 @@ namespace Tests.PlayMode
                 );
             
             yield return _tc.StartCoroutine(_tc.Run());
+            
+            LogResults("Files failed to upload, check logs for response",_tc.successCount==2);
         }
 
         [UnityTest]
@@ -391,7 +408,7 @@ namespace Tests.PlayMode
             
             yield return WaitForReturn(new[] { id1, GetUploadId(_tc.m_response) });
 
-            Assert.IsFalse(_failCount > 0);
+            LogResults("Files failed to upload, check logs for response", _tc.successCount==2);
         }
         
         private IEnumerator DeleteAllFiles()
@@ -404,7 +421,7 @@ namespace Tests.PlayMode
         {
             int count = 0;
             bool sw = true;
-            Debug.Log("Waiting for return...");
+            Debug.Log("Waiting for file to upload...");
             BrainCloudClient client = _tc.bcWrapper.Client;
 
             client.Update();
@@ -446,6 +463,7 @@ namespace Tests.PlayMode
         {
             _returnCount++;
             _failCount++;
+            _tc.m_statusMessage = jsonData;
         }
         
         private string GetCloudPath(Dictionary<string, object> response)
